+++ crates/objc2/src/topics/about_generated/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - `SystemConfiguration` / `objc2-system-configuration`.
 * Added new framework crates:
   - `Cinematic` / `objc2-cinematic`.
+  - `CoreSpotlight` / `objc2-core-spotlight`.
   - `ExecutionPolicy` / `objc2-execution-policy`.
   - `GameplayKit` / `objc2-gameplay-kit`.
   - `ImageCaptureCore` / `objc2-image-capture-core`.